+++ src/components/button/Button.tsx
@@ -2,18 +2,22 @@ import classNames from "classnames";
 import { UtilityStyledComponent } from "@uiComponents";
 import { NavLink } from "react-router-dom";
 import { ButtonComponentProps } from "@uiComponentTypes";
+import { forwardRef } from "react";
 
-const Button = ({
-  as: Component = "button",
-  display,
-  size = "md",
-  variant = "solid-primary",
-  href,
-  target,
-  type = "button",
-  to = "/404",
-  ...utilityProps
-}: ButtonComponentProps) => {
+const Button = (
+  {
+    as: Component = "button",
+    display,
+    size = "md",
+    variant = "solid-primary",
+    href,
+    target,
+    type = "button",
+    to = "/404",
+    ...utilityProps
+  }: ButtonComponentProps,
+  ref?: React.Ref<React.ElementRef<"button">>
+) => {
   const className = classNames(
     "btn",
     `btn-${size}`,
@@ -21,27 +25,20 @@ const Button = ({
     variant && "btn-" + variant
   );
 
-  const conditionalProps: {} = {
+  const conditionalProps = {
     ...utilityProps,
-    ...(Component === "anchorLink" && { href: href, target: target }),
+    ...(Component === "a" && { href: href, target: target }),
     ...(Component === "button" && { type: type }),
     ...(Component === "navLink" && { to: to }),
   };
 
-  if (Component === "button")
+  if (Component === "button" || Component === "a")
     return (
       <UtilityStyledComponent
-        as="button"
+        as={Component}
         {...conditionalProps}
         className={className}
-      />
-    );
-  if (Component === "anchorLink")
-    return (
-      <UtilityStyledComponent
-        as="a"
-        {...conditionalProps}
-        className={className}
+        ref={ref}
       />
     );
   if (Component === "navLink")
@@ -49,4 +46,4 @@ const Button = ({
   //return <UtilityStyledComponent as={NavLink} to={to} {...conditionalProps} className={className} />;
 };
 
-export default Button;
+export default forwardRef(Button);