+++ src/components/card/CardContent.tsx
@@ -1,9 +1,14 @@
 import { UtilityStyledComponent } from "@uiComponents";
 import { CardContentComponentProps } from "@uiComponentTypes";
+import { forwardRef } from "react";
 
-const CardContent = ({ ...utilityProps }: CardContentComponentProps) => {
+const CardContent = (
+  { ...utilityProps }: CardContentComponentProps,
+  ref?: React.Ref<HTMLDivElement>
+) => {
   return (
     <UtilityStyledComponent
+      ref={ref}
       as="div"
       className={"card__content"}
       {...utilityProps}
@@ -11,4 +16,4 @@ const CardContent = ({ ...utilityProps }: CardContentComponentProps) => {
   );
 };
 
-export default CardContent;
+export default forwardRef(CardContent);