+++ src/components/hover-card/Heading.tsx
@@ -1,13 +1,15 @@
 import { Typography } from "@uiComponents";
 import { HoverCardHeadingComponentProps } from "@uiComponentTypes";
 import useHoverCardProvider from "./useHoverCardProvider";
+import { forwardRef } from "react";
 
-const HoverCardHeading = ({
-  ...utilityProps
-}: HoverCardHeadingComponentProps) => {
+const HoverCardHeading = (
+  { ...utilityProps }: HoverCardHeadingComponentProps,
+  ref?: React.Ref<HTMLHeadingElement>
+) => {
   const isUsedInsideParentHoverCard = useHoverCardProvider();
   isUsedInsideParentHoverCard;
-  return <Typography as="h3" {...utilityProps} />;
+  return <Typography as="h3" {...utilityProps} ref={ref} />;
 };
 
-export default HoverCardHeading;
+export default forwardRef(HoverCardHeading);