+++ src/components/hover-card/OverlayHeading.tsx
@@ -1,14 +1,15 @@
 import { Typography } from "@uiComponents";
 import useHoverCardProvider from "./useHoverCardProvider";
 import { TypographyComponentProps } from "@uiComponentTypes";
+import React, { forwardRef } from "react";
 
-const OverlayHeading = ({
-  as = "h3",
-  ...utilityStyledComponent
-}: TypographyComponentProps) => {
+const OverlayHeading = <T extends React.ElementType>(
+  { as = "h3", ...utilityStyledComponent }: TypographyComponentProps,
+  ref?: React.Ref<React.ElementRef<T>>
+) => {
   const isUsedInsideParentHoverCard = useHoverCardProvider();
   isUsedInsideParentHoverCard;
-  return <Typography as={as} {...utilityStyledComponent} />;
+  return <Typography as={as} {...utilityStyledComponent} ref={ref} />;
 };
 
-export default OverlayHeading;
+export default forwardRef(OverlayHeading);