+++ src/components/layout/Column.tsx
@@ -1,8 +1,19 @@
 import { ColumnComponentProps } from "@uiComponentTypes";
 import { UtilityStyledComponent } from "@uiComponents";
+import { forwardRef } from "react";
 
-const Column = ({ ...utilityProps }: ColumnComponentProps) => {
-  return <UtilityStyledComponent as="div" {...utilityProps} className="col" />;
+const Column = (
+  { ...utilityProps }: ColumnComponentProps,
+  ref?: React.Ref<HTMLDivElement>
+) => {
+  return (
+    <UtilityStyledComponent
+      as="div"
+      {...utilityProps}
+      className="col"
+      ref={ref}
+    />
+  );
 };
 
-export default Column;
+export default forwardRef(Column);