+++ src/components/link/Link.tsx
@@ -1,8 +1,12 @@
 import { UtilityStyledComponent } from "@uiComponents";
+import { forwardRef } from "react";
 
 export type LinkComponentProps = {};
-const AsAnchorLink = ({ ...rest }: LinkComponentProps) => {
-  return <UtilityStyledComponent as="a" {...rest} />;
+const AsAnchorLink = (
+  { ...rest }: LinkComponentProps,
+  ref?: React.Ref<HTMLAnchorElement>
+) => {
+  return <UtilityStyledComponent as="a" {...rest} ref={ref} />;
 };
 
-export default AsAnchorLink;
+export default forwardRef(AsAnchorLink);