+++ src/types/Props/columnProps.ts
@@ -1,12 +1,23 @@
 export type ColumnOptions = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
 
-export type ColumnBreakpointProps = {
+export type ColumnBreakpointOptions = {
   xs?: ColumnOptions;
   sm?: ColumnOptions;
   md?: ColumnOptions;
   lg?: ColumnOptions;
   xl?: ColumnOptions;
   xxl?: ColumnOptions;
+};
+
+export type ColumnBreakpointProps = {
+  columnBreakpoints?: {
+    xs?: ColumnOptions;
+    sm?: ColumnOptions;
+    md?: ColumnOptions;
+    lg?: ColumnOptions;
+    xl?: ColumnOptions;
+    xxl?: ColumnOptions;
+  };
 };
 
 // export type ColumnBreakpointProps = {