+++ src/types/Props/gapProps.ts
@@ -1,0 +1,4 @@
+import { ValueAndUnitProps } from "types/shared";
+import { SizeOptions } from "./sizeProps";
+
+export type GapProps = { gap?: SizeOptions | ValueAndUnitProps };